Remove link to force 0 implementation
About forcing 0 before non 0 approve, the standard states "THOUGH The contract itself shouldn't enforce it, to allow backwards compatibility with contracts deployed before".
But in the implementation, it lists a token forcing the 0 for approve, which is a violation of the standard.
This is quite confusing. Therefore I propose to remove the link to implementations not following the standard.