Adding co-author
When Vitalik initially proposed the idea it was quite different, using the binary form of the address, which added complexity. I created the first implementation in that form, and suggested the simplified version `If it's 8 or upper (including a-f) then it should be uppercase, otherwise, it should be lowercase.` I also pushed the first implementations. His original post has been edited by EIP editors as we didn't have the PR format yet.

Would that be enough to be a co-author? I don't mind if this is rejected, just wanted to point that up.