Change name and symbol functions to view
The mutability of the `name()` and `symbol()` functions in the metadata extension should be `view` not `pure`. Having it set to `pure` assumes that the token name and symbol will be hard-coded into the contract functions, rather than set (for example, in the constructor).

The mutability guarantee rules allow you to go stricter, (ie move from `view` to `pure`), but not the other way, so dropping it down to `view` wouldn't take anything away but would allow for these variables to be set and stored elsewhere in the contract.