Updates the error response for not found block.
Invalid parameters isn't quite right, since the parameter in this case was valid, but the resource doesn't exist or isn't available.  Also changes `data` to just contain the block hash so that clients have an easier time parsing out the not-found block (no need to use a regex to pull out the block hash).