Fixed subscription API, and simplified
After carefully reading this again, i noticed that there was an issue with the subscription.
One should put the subscription type in, so that the notification can pre-filter based on the subscription type (`eth_subscription`, `ssh_subscription`). Athen the returned object is within this subscription type, and mapping of sub IDs can be done properly.

Also the standard should not talk about JSON RPC, as this provider is meant to abstract that away. Under the hood it probably will have to create a son roc object, depending on how it communicates with the node.

I'm personally not sure if the `accountsChanged` and `networkChanged` events should be in this standard.

I certainly think the `eth_requestAccounts` should be in another standards proposal and not in the initial one here.