eip-2047 - Temporary Data Persistence
Hi all,

We are building an accountable watching service, PISA, to watch and protect off-chain channels. Generally, it is important that on-chain challenges/disputes are temporarily stored on the blockchain such that it can be used by a client as indisputable proof that PISA has cheated and thus hold us financially accountable. 

The DataRegistry is an on-chain smart contract that ensures data is kept around for a minimal period of time, INTERVAL, such that it will eventually be discarded and not held in the network's state forever. It has a very simple API (setData, fetchRecord). 

In generally, while it is useful for PISA and off-chain protocols, it is actually useful to facilitate cross-contract communication and for building accountable watching services for most smart contracts. Thus we hope to propose it as a standard to let other companies/teams facilitate temporary data persistence.

Do let us know what you think of the EIP - we have provided an example implementation - so we would also be very keen to hear what people think about the code and whether it can be improved in whatever manner!