EIP-1193: Remove web3 compatibility section
This EIP had a section describing how to be compatible with a certain beta version of web3 1.0.0. This would look fairly arbitrary to the uninformed reader. Also, it would be even weirder in the near future, as some changes to web3's versioning will be announced soon.

This PR removes that section, making the EIP independent from any particular library consuming it.