EIP 1102: Loosens the constraints a bit.
I think overall this EIP is still too prescriptive in its wording, but I wanted to keep the initial change small to see if there is support for the direction first.

My view on standards is very libertarian: The purpose of standards isn't to tell people how to behave on an individual level, it is to tell people how to behave when they are interfacing with third parties so that communication is possible without first knowing everyone you might want to talk to.  Saying that the "standard language for signage is English" is useful, because it lets people know that if they want to make signs that anyone can read they should do so in English, and if people want to be able to read signs they should learn English.  Saying that "everyone must learn English and speak it in the home" on the other hand is overly prescriptive and doesn't allow for evolution and exploration of alternatives and it doesn't provide any meaningful value to society at large.

In the same vein, this standard should not prescribe how signers should behave, but it _should_ describe what the standard mechanism is for communicating between the browser and the dapp, particularly around accessing user addresses.  This change tries to move the needle closer to the latter, and away from the former.

I think this PR could be further improved by removing a _lot_ of the prescriptive text around what signers _should_ do and instead relegating the "consider prompting users" to the Rationale section.  I would like to wait to get feedback on the general changes before putting the time into that however.