Updated To Final
Notes for the call have not yet been published but, call 77 can be found on https://github.com/ethereum/pm

The condition is that block reward issuance is clarified that there is no change to the block reward. The last call review period has completed and so it is ready to be merged as Final.