Update eip-634.md
Hey Rick. 
Thanks for the good work. Hope to meet you in person one day.
I'm updating my dutch.eth name to include my Twitter handle and came across this description of what the vnd.twitter should be. It was a little bit confusing having the "SHOULD" capitalized and the "not" uncapitalized, so figured I'd suggest this small change and capitalize both words.
-Dutch