EIP-908: Marking as abandoned.
This EIP hasn't been touched for over 2 years.  As an author on the EIP, I don't even remember what it is at this point and after reviewing it I no longer hold any opinions I may have held on the topic two years ago.  Unless @jamesray1 plans on pushing for this, I am recommending we mark it as abandoned.

If someone later decides to pick it up again, we can always bring it back to DRAFT.