EIP-1559: Changes specification to functional Python code
This may be contentious, or it may not be, but I wanted to get a feel for other people's thoughts on just having code for the specification.  My reasoning for this change can be summed up nicely in the following comic, and I was finding myself frustrated trying to make sure that the specification was clear enough while using an arbitrary language that was being made up on the fly: https://www.commitstrip.com/wp-content/uploads/2016/08/Strip-Les-specs-cest-du-code-650-finalenglish.jpg

I did my best to try to include comments for all of the important bits, but I'm certainly open to adding more comments for the human readers.  This code should run out-of-the-box in Python 3.6+ and it should type check with MyPy (you'll need to install one pip package).  In a future PR, I would like to include test cases that can actually run against this code (I currently have two that I used just to validate that the code compiled and ran, but their format isn't very portable yet).  These test cases would, presumably, go in the Tests section of the EIP though rather than in the specification.

Please note: I have tried not to apply any normative changes to the EIP in this PR (such as removing the "miner pays fees" branch that has been discussed) as I would like to keep this PR focused purely on how the specification is written, not what it specifies.  However, there were some problems with the spec that had to be addressed before I could write valid code (part of the reasoning why I switched to a code spec instead of an English spec) but the python code **SHOULD** align with the current implementations.