Fix eip-1679 rendering (#2873)
The HTML generator is having trouble with the double colon in the title line of EIP-1679. I removed the quotes when adding the new validator, and for some reason the HTML generator raised it as a warning rather than an error. I've logged this issue with eipv, but we should go ahead and fix this ASAP.