EIP-1: Removes history and updated list.
Let the change control system do what the change control system was designed to do.  It is entirely redundant to store change history in the EIP itself and it just adds an ever growing clutter to an already too long (IMO) process description.  If anyone wants to see the history they easily can.  Also, the history section said to see the history button in the top right of this EIP but that only works if you view this EIP from GitHub, not from any other interface.