Add rationale for precompile; s/relayer/invoker/ [EIP-3074]
To hopefully clear up some of the confusion when discussing this EIP,
I've renamed what I originally referred to as a "relayer" to the
"invoker", which I think is more clear.

The established vocabulary seems to agree that relayer is actually the
actor who _sends_ the transaction, and conflating that with some
optional intermediary contract is just a bad idea.