Implementation => Reference Implementation (#3078)
In reality, the `Implementation` section almost always involves someone linking out to a personal or company repository when they create an EIP from the template.  External repositories cause problems because:
1. Links move/change over time and eventually the link will be dead.
2. The implementation likely isn't CC0 as the EIP is, and naive users may follow the link not realizing it is GPL and they are screwing themselves legally.
3. People are using EIPs for advertising their project, and they look for any opportunity they can to link out to their project in the body of the EIP.
4. People often link to some monolithic repository (e.g., go-ethereum) as the "implementation" which is less than useful.

Also, the description was written back when EIPs were part of the hard fork coordination process, so that has been rewritten.

This change makes it so it is more clear that this section is optional and that external links are not the preferred solution to completing this section.