1559: Changes starting base fee to 1 nanoeth and floors base fee at 0.
Original spec had the starting base fee (at fork block) as 1 nanoeth, but this was unintentionally lost along the way.  I believe current clients have implemented it the original way so this change shouldn't require any contention.

Also fixed a bug in the spec that allowed the base fee to go negative by just flooring it to 0.