3074: Removes change to ORIGIN.
As much as I *hate* ORIGIN with a burning passion and would love to have it broken/removed, I don't think sneaking it in with this EIP is the right way to go.  ORIGIN references the transaction that initiated the transaction, and that doesn't change by adding a new `CALL` opcode.  I would 100% support a separate EIP that removes `ORIGIN` or otherwise breaks it, just not embedded in a new opcode.