Update eip-2981.md
Proposed changes:
Remove/modify statements that imply that this is an extension of ERC-721.  There is no hard dependency.
Add a note to the effect that it is not exclusive to ERC-721.
Add requirement that marketplaces should not accept royalty amounts where the NFT uses the extra data, unless they are operating under an extended EIP.
Rename _INTERFACE_ID_ERC721ROYALTIES to _INTERFACE_ID_ERC2981, per the pattern used by ERC721.
Extra wording around the unenforcability of royalties to stress that even enforcing in code can be defeated.
Re-write the backward compatibility section.  I think this predated our cut-down version and there are no backward compatibility concerns other than 
- existing NFTs obviously cannot take advantage of it and 
- marketplaces will need to accommodate this EIP in addition to their current strategies.