Removes gas references.
Gas prices can (and frequently do) change, and no contract or specification can make any guarantees on how much an operation will cost in the future.  Saying "this will cost less than 30,000 gas" is an assertion that this specification **CANNOT** make.

This is a normative change to a final EIP, but I think in this case it is acceptable because the EIP currently makes an assertion about the universe that it cannot reasonably make and it should be fixed to no longer make such an assertion.