Delete EIP-5000
This appears to be authored by someone actively attacking the EIP numbering system.  Not only did they appear to number snipe, but they blatantly ignored editor number assignment and picked the number they wanted anyway.  It was merged, but I suspect the editor who approved it just missed the assignment comment because it was marked by GitHub as "resolved".

Timeline:
* At around 18 minutes after the hour, a spammer created issue #4998.  This has since been deleted as spam.
* At around 19 minutes after the hour, a spammer created issue #4999.  This has since been deleted as spam.
* At around 19 minutes after the hour, @hrkrshnn created PR #5000.
* An hour and 7 minutes later [Sam reviewed it](https://github.com/ethereum/EIPs/pull/5000#pullrequestreview-942583325) and assigned number 4998 with a comment indicating why 5000 wasn't assigned.
* ~18 days later (with no updates from the author) [lightclient reviewed it](https://github.com/ethereum/EIPs/pull/5000#pullrequestreview-960575707)
* ~Over two months later (still no updates from the author) [lightclient reviewed it again](https://github.com/ethereum/EIPs/pull/5000#pullrequestreview-1037337130)
* About an hour after that the author made the [first change](https://github.com/ethereum/EIPs/pull/5000/commits/edaa3fb55d681b271f3a0841a0d16f61e8e8fe53) to the PR, in which they ignored Sam's assigned number and took 5000 anyway.
* Review proceeded from there as normal and was eventually merged.

----

We have a serious spam problem on this repository and it is a huge editor time sink.  At best, the author of this PR wasn't the spammer but they were running a number squatting bot and they blatantly ignored editor number assignment.  At worst they are also the spammer and part of a serious problem.

I propose we delete this EIP and either ban the author from the repository for this behavior.  The blatant disregard for an editor's number assignment without discussion combined with the blatant number sniping makes me classify this as an active attack.

----

Also, since an editor is a co-author on this EIP I think this case should be taken *far* more seriously than normal.  Editors should be held to a far higher standard than regular drive-by authors and while I am not suggesting @axic played an active role in this attack (they may not have known about it), I think that a strong show of intolerance against malicious behavior here is necessary due to this association as we cannot be seen to be giving preferential treatment to editors.