eip-5202: rename factory to blueprint
It was pointed out to me in private discussion that "factory" is an
overloaded term. In the Maker system, "factory" refers to a contract
which deploys other contracts. In object-oriented programming, a
"factory" is an object for creating other objects. Based on this, the
term "blueprint" is used to avoid confusion with these other usages of
"factory".