Removes MicahZoltu as EIP editor.
Context: We now have other editors who are relatively active so the response time for authors isn't as bad as it used to be (a week vs a year).  It also sounds like the process for consensus changes will be moving over to the executable specs repository in the near future and since that is the piece I care most about, I am less concerned now than I used to be about the EIP process remaining credibly neutral.