Synchronize EIP-4844 cryptography with consensus specs
This commit introduces various fixes and improvements to the KZG-related code of EIP-4844. This commit brings the EIP
code up-to-speed with the consensus-specs code.

Note that instead of updating the cryptographic functions we instead link to the relevant parts of the consensus-specs
repository. This is done to avoid code duplication between the two repositories -- an approach that seems prudent in
cross-execution-and-consensus situations like EIP-4844.