Assign EIP number
EIP number 6059 was assigned as this is the number of the PR to add this
proposal.

This was doen before being assigned the EIP number by one of the editors
based on the common practice of assigning EIP numbers based on the PR
number at which the proposal was added. In case the number of the
proposal should be different, the number can easiliy be changed.