Change name to "Lockable tokens"
When this standard was created, the term "Soulbound tokens" was cool and new. Many people understood that a Soulbound token is "bound to a soul" and so it is a permanently locked token of an account. But since then, it has become clear that a Soulbound token is an umbrella term for all kinds of tokens that break conventional norms around private property assumptions. And besides, EIP-5192 implements dynamic locking of a token. So "Lockable Tokens" is a much better term for this specification and also for developers arriving at it in the future.

I'm hence asking the EIP editors to make an exception regarding changing the contents of this EIP. The interface has not changed. Developers can still rely on using this token standard. The only change is an improvement in terms of accuracy in identifying what it is: A lockable token.