Update eip-1328.md
Providing a relay URL in the WC URI enables the application to indicate what relay they should use. 
Currently WC20 clients can choose to specify a custom relay. However in practice they are forced to use relay.walletconnect.com which gives stats (IP, User-Agents, Application ID) about the users to the central relay and is not acceptable for many applications.