normative language adjustments and additional privacy notes
This modifies normative language to align it better with the expected
outcomes defined. Additionally, it defines how pairwise SIWE accounts
MAY be used (there's a tradeoff here that asset accounts will probably
still act as global cross site tracking identifiers).

One other notable aspect is that "domain" was changed to "origin" in
order to make sure that the scheme is also included. This is important
to prevent attacks where the same site is presented over multiple protocols
such as wss:// and https:// which would potentially allow for a replay
attack to occur to the same backend service. If these services weren't
treated as independent and shared a backend service, this could potentially
allow for multiple authenticated sessions to be created even if the
nonce was implemented properly.

other just general clean up was also done in this EIP in order to
prepare it for final call.