Add EIP: Token Mapping Slot Retrieval Extension (#6823)
This proposal suggests an extension to the ERC-20/ERC-721/ERC-1155 standards
that allows retrieving the reserved storage slot of the mapping responsible
for tracking ownership of compliant tokens. The proposed extension enables
accurate off-chain simulations, eliminates the reliance on events, and
enhances the precision of data access from storage. The proposed extension
is a single function named getTokenLocationRoot() that returns the reserved
storage slot for the mapping type in any ERC-20/ERC-721/ERC-1155 compliant
smart-contract implementation. The proposed extension does not have any additional
cost for the caller, and there is a negligible increase in the deployment cost
of the contract.