Update EIP-6381 to be a public good repository
After a discussion about EIP-6381 in the discussion thread, we concluded
that it is best to rewrite the proposal from an ERC-721 extension to a
public good repository separate from the token contracts. This way any
non-fungible token can be emoted at without the need for the token
interface to include this EIP. Additionally, the EIP became compatible
with 1155 in addition to 721.