Updates interface to match interface Id
During the process there were some changes that we failed to update in the main document.
These changes are actually in place on the interface id which the document includes.

The main problem is the nestTransferFrom method was missing the data parameter. This
was missing only in the document but not on the sample implementation.

The ChildTransferred event was missing the toZero event on both sample implementation
and interface. Although this does not affect the interface, it is a very useful parameter
to detect if a child was abandoned.

We took the change to relocate the Child struct, so both structs are together before events.