Add Dynamic On-Chain Token Attributes Repository EIP
This proposal introduces the ability to assign attributes to NFTs in a
public non-gated repository smart contract that is accessible at the
same address in all of the networks. The repository smart contract is
designed to be a common-good repository, meaning that it can be used by
any ERC-721 or ERC-1155 compatible token.