Update EIP-7543: Add "arbitrary precision" to title
it actually is arbitrary precision. the theory is. i did not add that before because in practice, the machine is limited. however, i think it is important to understand, that it has as much precision as any finite machine allows. within that limit, you can count on arbitrary precision here.