EIP-6110: change request to flat encoding
I am proposing to change the encoding of deposit requests to be a flat concatenation of
the fields returned by the contract. The extra layer of RLP encoding is not necessary, and
by removing it, we can avoid defining the structure of these requests in the execution
layer client implementation.