EIP-7251: change request to flat encoding
I am proposing to change the encoding of consolidation requests to be exactly equal to how
they are returned by the contract. The extra layer of RLP encoding is not necessary, and
by removing it, we can avoid defining the structure of these requests in the execution
layer client implementation.