Update EIP-7623: Clarify the gas refunds handling
Previously, the `evm_gas_used` (now renamed to `execution_gas_used`)
was not specified. This lead to two different interpretations whenever
the `execution_gas_used` was the value after or before applying refunds.

We argue that the interpretation that `execution_gas_used` don't
have refunds subtracted is incorrect because it would make the "current"
formula for the `tx.gasUsed` incomplete and confusing as the total gas
used by a transaction would actually be `tx.gasUsed - refunds`.

Using this reasoning, this change clarifies the specification
by following the interpretation that the `execution_gas_used` is
the value after the refunds are applied.