Add EIP: eth/vhash - make mempool messaging vhash aware
This EIP eliminates the need to redistribute blob content in the mempool if
only the metadata (fees) of a transaction are updated, making RBF
(replace-by-fee) more efficient and cheaper for the network. It achieves this
modifying the devp2p ‘eth’ protocol to address blobs in type3 transaction
sidecars by content (vhash).

Signed-off-by: Csaba Kiraly <csaba.kiraly@gmail.com>